====================================== TYPE: bug fix
KEYWORDS: RSL NCEP_DEBUG_MULTIDIR cpp CLFAGS

SOURCE: Mark Filipak (University of Edinburgh)

PURPOSE: Fix compiler error when using the -DNCEP_DEBUG_MULTIDIR option

DESCRIPTION OF CHANGES: 
The option NCEP_DEBUG_MULTIDIR causes each of the rsl.{out,error}.* files to be placed in a separate directory (for example, rsl.out.0000 and rsl.error.0000 both go in a single directory, and no other rsl file is in that directory).  This can be convenient if there are a large number of those files.  When that option is activated (as a cpp flag), several directories are checked for existence (and attempts are made to create them if the directories do not exist).  When there are problems with directory existence, the code tries to help out the user and return a helpful error message.  This is handled through the errno.h file.  This file is only ever needed when NCEP_DEBUG_MULTIDIR is selected, and this file errno.h is always needed when NCEP_DEBUG_MULTIDIR is selected.  It is a compile-time error to not have the errno.h file when required.

c_code.c(117): error: identifier "errno" is undefined
          if ( rc != 0 && errno==EEXIST) rc=0;

The easy thing to do is to test whether NCEP_DEBUG_MULTIDIR is defined, and then #include the required file towards the top of c_code.c

#ifdef NCEP_DEBUG_MULTIDIR
#  include <errno.h>
#endif

LIST OF MODIFIED FILES :
M       external/RSL_LITE/c_code.c

TESTS CONDUCTED: 
Without this change, when the -DNCEP_DEBUG_MULTIDIR is added to CFLAGS, the WRF code does not compile.  After the #include code is added to c_code.c, the WRF code now compiles.  Ming Chen tested the option: when working correctly, you do indeed get a separate directory, which has subdirectories, each of which holds a single rsl.out.* plus a single rsl.error.* file.
======================================


git-svn-id: https://svn-wrf-model.cgd.ucar.edu/branches/RELEASE_BRANCH_371@8580 b0b5d27b-6f0f-0410-a2a3-cb1e977edc3d