Branch commit.  Fixed problem with quilt decomposition found by Sam on Zeus.
It showed up on Zeus when Sam had to use a larger number of quilt tasks because of
lower per-node memory limits on that system.  Prior to this, the number of processors
in Y and the number of tasks (nio_tasks_per_group) happily avoided a pathological
combination.  Lucky Sam.



git-svn-id: https://svn-wrf-model.cgd.ucar.edu/branches/jm_merge_bHWRF@8589 b0b5d27b-6f0f-0410-a2a3-cb1e977edc3d